separate inline-editor module
