+++ react-app/src/inline-editor/actions/index.js
