+++ react-app/src/inline-editor/components/App.js
