+++ react-app/src/inline-editor/reducers/cursor.js
