+++ react-app/src/inline-editor/reducers/index.js
