+++ react-app/src/inline-editor/reducers/lines.js
