+++ react-app/src/inline-editor/render/mermaid.js
