+++ react-app/src/inline-editor/utils/inlineDecorator.js
