+++ react-app/src/inline-editor/utils/render.js
